# whatsappclone/src/main/java/com/myapp/whatsappclone/common/StringResponse.java
package com.myapp.whatsappclone.common;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class StringResponse {

    private String response;
}